Updated project settings for Xcode 6.1

+++ ZKCDTrailer.m
@@ -105,7 +105,7 @@
 }
 
 - (NSData *) data {
-	NSMutableData *data = [NSMutableData zk_dataWithLittleInt32:self.magicNumber];
+	NSMutableData *data = [NSMutableData zk_dataWithLittleInt32:(UInt32)self.magicNumber];
 	[data zk_appendLittleInt16:self.thisDiskNumber];
 	[data zk_appendLittleInt16:self.diskNumberWithStartOfCentralDirectory];
 	[data zk_appendLittleInt16:self.numberOfCentralDirectoryEntriesOnThisDisk];
@@ -114,8 +114,8 @@
 		[data zk_appendLittleInt32:0xFFFFFFFF];
 		[data zk_appendLittleInt32:0xFFFFFFFF];
 	} else {
-		[data zk_appendLittleInt32:self.sizeOfCentralDirectory];
-		[data zk_appendLittleInt32:self.offsetOfStartOfCentralDirectory];
+		[data zk_appendLittleInt32:(UInt32)self.sizeOfCentralDirectory];
+		[data zk_appendLittleInt32:(UInt32)self.offsetOfStartOfCentralDirectory];
 	}
 	[data zk_appendLittleInt16:[self.comment zk_precomposedUTF8Length]];
 	[data zk_appendPrecomposedUTF8String:self.comment];
